Update package and refactor backend context to resolve cycle warning.
