+++ app/react-native-app/package.json
@@ -29,9 +29,9 @@
     "moment": "^2.29.4",
     "pullstate": "^1.25.0",
     "react": "18.2.0",
-    "react-native": "0.71.6",
+    "react-native": "0.71.14",
     "react-native-dotenv": "^3.4.8",
-    "react-native-gesture-handler": "^2.9.0",
+    "react-native-gesture-handler": "~2.9.0",
     "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-modal-datetime-picker": "^14.0.1",
     "react-native-reanimated": "~2.14.4",
